qcacmn: Change WMI API's event_buf length type to uint32_t
Change all the WMI API's event buf length param to
uint32_t to adjust to event buf length param type
in the wmi_unified_event_handler().These changes are
needed to converge on the common service ready handlers.

CRs-Fixed: 1110768

+++ wmi/inc/wmi_unified_api.h
@@ -1134,7 +1134,7 @@ QDF_STATUS wmi_extract_fw_abi_version(void *wmi_hdl,
 QDF_STATUS wmi_check_and_update_fw_version(void *wmi_hdl, void *ev);
 
 uint8_t *wmi_extract_dbglog_data_len(void *wmi_hdl,
-				void *evt_b, uint16_t *len);
+				void *evt_b, uint32_t *len);
 
 QDF_STATUS wmi_send_ext_resource_config(void *wmi_hdl,
 				wmi_host_ext_resource_config *ext_cfg);
